requirements.txt
# Discord Bot Framework
py-cord==2.4.1

# Database
mysql-connector-python==8.2.0

# Environment Variables
python-dotenv==1.0.0

# Image Processing (for rank cards)
Pillow==10.1.0